Move tbtc interface to the tbtc.ts file
Main changes:
- move tbtc interface to the tbtc.ts file and remove the tbtc.interface.ts file
completely. We want to keep interface and class in the same file since it's
related.
- add comments to the interface methods
- change the occurences of `deposit` variables with a type of
`DepositScriptParameters` to `depositScriptParameters`
