+++ src/threshold-ts/tbtc/index.ts
@@ -1,2 +1,1 @@
 export * from "./tbtc"
-export * from "./tbtc.interface"
