+++ src/threshold-ts/tbtc/tbtc.ts
@@ -18,12 +18,72 @@ import {
   decodeBitcoinAddress,
   UnspentTransactionOutput,
 } from "@keep-network/tbtc-v2.ts/dist/bitcoin"
-import { ITBTC } from "./tbtc.interface"
 import { ElectrumClient, EthereumBridge } from "@keep-network/tbtc-v2.ts"
 import BridgeArtifact from "@keep-network/tbtc-v2/artifacts/Bridge.json"
 import { BitcoinConfig, BitcoinNetwork, EthereumConfig } from "../types"
 import TBTCVault from "../../../node_modules/@keep-network/tbtc-v2/artifacts/TBTCVault.json"
 import { Contract } from "ethers"
+
+export interface ITBTC {
+  /**
+   * Gets Bitcoin network specified in the bitcoin config that we pass to the
+   * threshold lib
+   * @returns {BitcoinNetwork}
+   */
+  getBitcoinNetwork(): BitcoinNetwork
+
+  /**
+   * Suggests a wallet that should be used as the deposit target at the given
+   * moment.
+   * @returns Compressed (33 bytes long with 02 or 03 prefix) public key of
+   * the wallet.
+   */
+  suggestDepositWallet(): Promise<string | undefined>
+
+  /**
+   * Creates parameters needed to construct a deposit address from the data that
+   * we gather from the user, which are eth address and btc recovery address.
+   * @param ethAddress Eth address in which the user will receive tBTC (it
+   * should be the address that the user is connected to).
+   * @param btcRecoveryAddress The bitcoin address in which the user will
+   * receive the bitcoin back in case something goes wrong.
+   * @returns All deposit script parameters needed to create a deposit address.
+   */
+  createDepositScriptParameters(
+    ethAddress: string,
+    btcRecoveryAddress: string
+  ): Promise<DepositScriptParameters>
+
+  /**
+   * Calculates the deposit address from the deposit script parameters
+   * @param depositScriptParameters Deposit script parameters. You can get them
+   * from @see{createDepositScriptParameters} method
+   * @returns Deposit address
+   */
+  calculateDepositAddress(
+    depositScriptParameters: DepositScriptParameters
+  ): Promise<string>
+
+  /**
+   * Finds all unspent transaction outputs (UTXOs) for a given Bitcoin address.
+   * @param address - Bitcoin address UTXOs should be determined for.
+   * @returns List of UTXOs.
+   */
+  findAllUnspentTransactionOutputs(
+    address: string
+  ): Promise<UnspentTransactionOutput[]>
+
+  /**
+   * Reveals the given deposit to the on-chain Bridge contract.
+   * @param utxo Deposit UTXO of the revealed deposit
+   * @param depositScriptParameters Deposit script parameters. You can get them
+   * from @see{createDepositScriptParameters} method
+   */
+  revealDeposit(
+    utxo: UnspentTransactionOutput,
+    depositScriptParameters: DepositScriptParameters
+  ): Promise<string>
+}
 
 export class TBTC implements ITBTC {
   private _bridge: EthereumBridge
@@ -60,6 +120,11 @@ export class TBTC implements ITBTC {
     this._bitcoinConfig = bitcoinConfig
   }
 
+  /**
+   * Returns the network that is currently used for bitcoin, but instead of
+   * "mainnet" it returns "main". It is specific to the tbtc-v2.ts lib.
+   * @return {string} Network that can be used in tbtc-v2.ts lib
+   */
   private _getBitcoinNetworkForTBTCLib = (): string => {
     const { network } = this._bitcoinConfig
     if (network === BitcoinNetwork.mainnet) {
@@ -135,11 +200,11 @@ export class TBTC implements ITBTC {
 
   revealDeposit = async (
     utxo: UnspentTransactionOutput,
-    deposit: DepositScriptParameters
+    depositScriptParameters: DepositScriptParameters
   ): Promise<string> => {
     return await tBTCRevealDeposit(
       utxo,
-      deposit,
+      depositScriptParameters,
       this._bitcoinClient,
       this._bridge,
       {
